requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.13.1
idna==3.3
MouseInfo==0.1.3
Pillow==8.4.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyJWT==2.3.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
twilio==7.7.1
urllib3==1.26.7
